// src/app/model/product.ts
export class products{
    productId : number;
    productName: string;
    productDescription: string;
    productStatus: string;
    productPrice: string;
    productImage: string;
    productLevel: string;
    gst: string;
    threshHold: string;
    productBasePrice: string;
    productType: string;
  }
  